LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-paused-animations-expected.txt


FAIL Test that the scroll animation is paused assert_equals: 'actual' unit type must be 'percent' for "timeline time reset" expected (string) "percent" but got (undefined) undefined
FAIL Test that the scroll animation is paused by updating animation-play-state assert_equals: 'actual' unit type must be 'percent' for "timeline time reset" expected (string) "percent" but got (undefined) undefined
